Fixed exception thrown from CatSpawner
I added a field to CatSpawner so you can just drag the button into it. It couldn't find it since it isn't a child of the cat spawner, so the variarble stayed null and Update() crashed.

I moved CatSpawner onto its own gameobject

Removed unnecessary extra copy of the script that was on the Start Wave button