Implemented camera panning
It supports WASD, arrow keys, and gamepad right stick. This can be changed in the InputActions asset in the Assets folder.

I made it use Cinemachine, though I ended up not getting much advantage out of it for now, but oh well. We can have nicer camera movement later with it.