Implemented cat nagivation
I ended up using the way point style, as it is a lot easier to configure on the map than a bunch of tiles that each have to be rotated the right way. The WayPoint class draws gizmos showing paths to all WayPoints specified in its NextWayPoints property. This way you can see how they are connected to each other in the editor! I put an arrow head in the center of each path line to indicate direction.

The cats' distractionThreshold and damageToPlayer properties are now exposed in the Inspector, so they can be easily tweaked. This is defaulted to 2 meters, because the ground is thick, and the WayPoints themselves are sitting at 0 on the Y-axis.

I added an arrivedAtWayPointDistance setting as well. The cat must be within this distance to be considered as having arrived at the next WayPoint, at which point it will target a new one. If the current WayPoint has more than one next point set in the Inspector, then one is simply chosen at random.

One last major change is to the cat classes themselves. I made a new base class, as like 99% of the code was duplicate. I kept the three cat classes, but they are mostly empty now and all inherit from CatBase. This should be a much nicer setup!