Volume Sliders for Music and Sound Effects
I had trouble getting the sound effect to not sound on awake despite it being deselected but other than that the volume slider is working. I could not get the meow sound effect to only sound after the user finished changing the value instead of while the slider was moving so it's a bit delayed to avoid the sound issue. I also added a button to get back to Main.