Made camera changes
I added a new setting on the CameraController that lets you set an object that it will be focused on when the level starts. That way we can easily control where the camera looks when a level begins. It starts at the starting area on Level1 now.

I added bars above the cats to show how distracted they are. CatBase has a static bool that controls whether or not they are displayed. That way it can easily be optional for players later if we want.

I also changed the tower base select color to dark blue.