Fixed laser towers
Added a circle at the end of the laser (just a quad with a transparent texture I made)

Made laser sweep back and forth across the path. The Laser tower has 3 new settings that let you specified how wide the laser's sweep back and forth is, how far ahead of cats it will target, and how long it takes the laser to sweep back and forth across the path once.