Added tower info popups to level1
The info displayed in the tower popups is stored in the scriptable objects in the Resources folder. They are automatically loaded at runtime by TowerInfoCollection, which also updates the cost field by pulling it from the tower prefabs. This way we don't need to update cost in the scriptable objects every time we change it in the tower prefabs.

I also made it so Tower's refund value is now refund percentage since Tiago said to have them all refund 85%. This can be changed in the inspector for each tower prefab.