Refactored laser tower to spec
I made a static class called WayPointUtils. You can call WayPointUtils.CompareWayPointPositions() to compare the position of two waypoints. it will tell you if a is before b, after b, if a and b are the same waypoint, or if there is no path connecting them. It just returns a simple enum value. I needed this to be able to figure out when a cat is approaching the path branch so the laser tower would know when to change the cat's next node to distract it onto the specified path.