Integrated FSM into towers
For now, this just adds a couple functions to Tower.cs that enable/disable the tower's target detection by enabling/disabling the collider. I also added a property to be able to easily check this state.

All towers use the base setup for now that is defined in Tower.InitStateMachine(). Laser tower is the only one currently overriding that function, as its condition for entering active state is slightly different from the others.