FIX: Crash when selecting tower while another is selected sometimes
+ Fixed another bug where the tower manipulation bar didn't update the properties panel when you select another tower.

As for the crash sometimes happening when you select a tower while another is selected, I wasn't able to figure it out. However, it is caused at line 97 in TowerBase.cs. I got it to happen twice in the first few minutes and have not been able to get it to happen again since.

I did notice that at runtime, one towerbase had its TowerSelectorUI field set to null. However, when I returned to the editor to fix it, it was set correctly! :O

So I used Find All References to see if it was being set to null somewhere, but the only references are in TowerBase.cs and none of those lines change the value of towerSelectorUI.

So yea, this is a tricky one. I guess we'll see if it happens again.