This is another update that comes from CELL
+ Added ReturnToPreviousState() method

The state machine now remembers the last state it was in so it can easily jump back to it.

It has an automatic bool param that defaults to true. If you set it to false, it will skip calling the previous state's OnEnter() method. This allows you to return to the previous state without having it reinitialize itself, as I did for the settings window in CELL. Otherwise it would reload the scene/reset the level.